# infra/openapi.yaml
openapi: 3.0.3
info:
  title: mp4-to-word-new backend API
  version: "1.0.0"
  description: |
    全量后端 OpenAPI 规范。Flask Blueprint 拆分：
      - `public_bp`        静态资源 / 根路径
      - `system_bp`        系统状态
      - `self_selected_bp` 自选股 + 自选分类 CRUD
      - `stock_chart_bp`   K 线 / 标的 / 题材 / 应用面分析 / 调度
      - `f10_bp`           F10 基础资料 / 题材 / 涨跌停 / 换手率
      - `mp4_history_bp`   MP4 历史记录
      - `transcription_bp` 视频转写工作流
      - `scheduler_bp`     调度器管理（手动 trigger / enable / disable / start / stop）

    直接把本文件拖入 Apifox 即可生成集合。
servers:
  - url: http://localhost:5000
    description: 本地开发（Flask 默认端口）
  - url: http://localhost:5173
    description: 通过 Vite dev server 代理
...
paths:
...
  /api/stock-chart/f10/topics:
    get:
      tags: [F10]
      summary: 概念 / 题材列表
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/stock-chart/f10/topic-compare:
    get:
      tags: [F10]
      summary: 题材对比（同题材其它票）
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: topic_id
          required: true
          schema: { type: string }
        - in: query
          name: section
          schema: { type: string, default: gndbzfsj }
        - in: query
          name: sort_by
          schema: { type: string, default: zdf }
      responses:
        "200":
          description: OK
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/stock-chart/f10/topic-stocks:
    get:
      tags: [F10]
      summary: 题材内成分股表（eltdx helper 风格）
      description: |
        对应 eltdx `client.helpers.topic_stocks(...)`。
        至少传 `seed_code` + (`topic_id` 或 `topic_name`)；
        不传 `topic_id` 时，会在 `seed_code` 关联的题材里按名称模糊匹配。
        `sort_by` 支持：
          - `zdf` 当日涨跌幅（默认）
          - `zdf_3d` / `zdf_5d` / `zdf_20d` / `zdf_60d` 近 3 / 5 / 20 / 60 日
          - `zdf_ys` 年初至今
      parameters:
        - in: query
          name: seed_code
          required: true
          schema: { type: string, default: "000001" }
        - in: query
          name: topic_id
          schema: { type: string }
        - in: query
          name: topic_name
          schema: { type: string, description: "不传 topic_id 时按名称匹配" }
        - in: query
          name: sort_by
          schema: { type: string, default: zdf }
        - in: query
          name: section
          schema: { type: string, default: gndbzfsj }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TopicStockTable" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/stock-chart/f10/stock-topics:
    get:
      tags: [F10]
      summary: 个股关联题材（eltdx helper 风格）
      description: |
        对应 eltdx `client.helpers.stock_topics(...)`。
        合并 `topic_ids` + `hot_topics`，带可读 `topic_name` / `relation_level` /
        `selected_date` / `reason` / `detail_id`。
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StockTopicsResponse" }
  /api/stock-chart/eltdx/industry-index-kline:
    get:
      tags: [F10]
      summary: 申万行业指数 K 线（完整历史）
      description: |
        通过 eltdx `client.bars.get(kind="index")` 拉取某个**申万行业指数**的
        完整 K 线历史。`code` 必须是 `sh8803XX` 格式，可选列表见
        `GET /api/stock-chart/eltdx/index-codes?kind=industry`。
        `period` 支持 `1m / 5m / 15m / 30m / 60m / 1d / 1w / 1M`。
        涨跌幅用本地算：`pct = (close - prev_close) / prev_close`。
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string, default: "sh880301" }
        - in: query
          name: period
          schema: { type: string, default: day }
        - in: query
          name: count
          schema: { type: integer, default: 120, maximum: 2000, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IndexKlineResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/stock-chart/eltdx/concept-index-kline:
    get:
      tags: [F10]
      summary: 概念主题指数 K 线（完整历史）
      description: |
        通过 eltdx `client.bars.get(kind="index")` 拉取某个**概念主题指数**的
        完整 K 线历史。`code` 必须是 `sh8804XX` 格式，可选列表见
        `GET /api/stock-chart/eltdx/index-codes?kind=concept`。
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string, default: "sh880401" }
        - in: query
          name: period
          schema: { type: string, default: day }
        - in: query
          name: count
          schema: { type: integer, default: 120, maximum: 2000, minimum: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IndexKlineResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/stock-chart/eltdx/index-codes:
    get:
      tags: [F10]
      summary: 行业 / 概念 指数代码全集
      description: |
        返回前端可下拉选的全部指数代码 + 中文名。
        数据源是项目内置的 `backend/services/stock/f10/index_codes.py`，
        全程不联网（开箱即用），eltdx 仅用于取行情。
      parameters:
        - in: query
          name: kind
          schema: { type: string, enum: [industry, concept, all, ""], default: "" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/IndexCodeItem" }
                  count: { type: integer }
                  source: { type: string, example: "index_codes.py" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/stock-chart/f10/theme-market:
    get:
      tags: [F10]
      summary: 主题市场行情
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: req_id
          schema: { type: string, default: "200743" }
      responses:
        "200": { description: OK }
  /api/stock-chart/sectors-market:
    get:
      tags: [F10]
      summary: 板块市场（统一分发器）
      description: |
        `category` 支持：
          - 数字 ID（默认 6 = 沪深A股）
          - 别名：`沪深A股` / `A股` / `all` / `全部`
          - 行业指数：`行业指数` / `industry`
          - 概念指数：`概念指数` / `concept`
      parameters:
        - in: query
          name: category
          schema: { type: string, default: "6" }
        - in: query
          name: sort_by
          schema: { type: string, default: 涨幅 }
        - in: query
          name: count
          schema: { type: integer, default: 100 }
        - in: query
          name: start
          schema: { type: integer, default: 0 }
        - in: query
          name: ascending
          schema: { type: string, enum: ["true", "false", "1", "0"], default: false }
      responses:
        "200": { description: OK }
  /api/stock-chart/sectors-market/industry:
    get:
      tags: [F10]
      summary: 32 个申万行业指数
      parameters:
        - in: query
          name: sort_by
          schema: { type: string, default: 涨幅 }
        - in: query
          name: count
          schema: { type: integer, default: 100 }
        - in: query
          name: start
          schema: { type: integer, default: 0 }
        - in: query
          name: ascending
          schema: { type: string, enum: ["true", "false", "1", "0"], default: false }
      responses:
        "200": { description: OK }
  /api/stock-chart/sectors-market/concept:
    get:
      tags: [F10]
      summary: 50 个概念主题指数
      parameters:
        - in: query
          name: sort_by
          schema: { type: string, default: 涨幅 }
        - in: query
          name: count
          schema: { type: integer, default: 100 }
        - in: query
          name: start
          schema: { type: integer, default: 0 }
        - in: query
          name: ascending
          schema: { type: string, enum: ["true", "false", "1", "0"], default: false }
      responses:
        "200": { description: OK }
...
  /api/stock-chart/limit-count/refresh:
    post:
      tags: [F10]
      summary: 强制刷新涨跌停统计
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                category: { type: string, default: 沪深A股 }
                max_pages: { type: integer, default: 80 }
      responses:
        "200": { description: OK }
  /api/stock-chart/turnover:
    get:
      tags: [F10]
      summary: 读某只票的换手率快照（不触发计算）
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: target_type
          schema: { type: string, default: stock }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TurnoverResponse" }
  /api/stock-chart/turnover/refresh:
    post:
      tags: [F10]
      summary: 刷新单只票的换手率
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol]
              properties:
                symbol: { type: string }
                target_type: { type: string, default: stock }
                period: { type: string, default: 1d }
                adjust: { type: string, default: qfq }
      responses:
        "200": { description: OK }
...
  /api/stock-chart/f10/stock-info:
    get:
      tags: [F10]
      summary: 股票信息
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/company-profile:
    get:
      tags: [F10]
      summary: 公司概况
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: section
          schema: { type: string, default: "8" }
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/business-composition:
    get:
      tags: [F10]
      summary: 主营业务构成
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: report_date
          schema: { type: string, format: date, nullable: true }
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/valuation:
    get:
      tags: [F10]
      summary: 估值
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: req_id
          schema: { type: string, default: "200191" }
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/finance-report:
    get:
      tags: [F10]
      summary: 财务报表
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: report_type
          schema: { type: string, default: zcfzb, description: "zcfzb/lrb/xjllb" }
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/finance-diagnosis:
    get:
      tags: [F10]
      summary: 财报诊断
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: section
          schema: { type: string, default: yynl }
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/stock-score:
    get:
      tags: [F10]
      summary: 评分
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: section
          schema: { type: string, default: pf }
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/profit-forecast:
    get:
      tags: [F10]
      summary: 盈利预测
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/ranking-detail:
    get:
      tags: [F10]
      summary: 排名详情
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: section
          schema: { type: string, default: scpmdela }
      responses:
        "200": { description: OK }
  /api/stock-chart/f10/governance:
    get:
      tags: [F10]
      summary: 公司治理
      parameters:
        - $ref: "#/components/parameters/QuerySymbol"
        - in: query
          name: section
          schema: { type: string, default: wgcl }
      responses:
        "200": { description: OK }

  # ====================================================================
  # MP4 History
  # ====================================================================
...
components:
  parameters:
...
    QuerySymbol:
      in: query
      name: symbol
      required: false
      description: |
        标的代码（`stock` 类型）或板块 / 指数代码。
        后端 helper `_symbol_arg()` 的默认值是 `000001`（**上证综合指数**），
        所以不传 symbol 时接口会返回上证指数的题材 / F10 / 估值。
        前端调用建议**显式传**，避免误以为是「随便一只票」的数据。
      schema:
        type: string
        default: "000001"
        examples:
          - "000001"
          - "600519"
          - "300750"
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              error_type: { type: string, example: bad_request }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
  schemas:
...
    TurnoverResponse:
      type: object
      properties:
        has_cache: { type: boolean }
        symbol: { type: string }
        target_type: { type: string }
        turnover:
          type: object
          additionalProperties: true

    # eltdx helpers 返回模型
    TopicStockRow:
      type: object
      properties:
        full_code: { type: string, example: "sz300975" }
        exchange: { type: string, example: "sz" }
        market_id: { type: string, example: "z" }
        code: { type: string, example: "300975" }
        name: { type: string, example: "商络电子" }
        rank: { type: integer, example: 1 }
        change_pct: { type: number, format: float, example: 19.97 }
        change_pct_3d: { type: number, format: float, example: 37.24 }
        change_pct_5d: { type: number, format: float, nullable: true }
        change_pct_20d: { type: number, format: float, nullable: true }
        change_pct_60d: { type: number, format: float, nullable: true }
        change_pct_ytd: { type: number, format: float, nullable: true }
        trading_date: { type: string, example: "20260528" }
        raw:
          type: object
          additionalProperties: true
    TopicStockTable:
      type: object
      description: "eltdx helpers.topic_stocks 返回模型"
      properties:
        seed_code: { type: string }
        topic_id: { type: string }
        topic_name: { type: string }
        sort_by: { type: string }
        section: { type: string }
        count: { type: integer }
        source: { type: string, example: "eltdx" }
        fetched_at: { type: string, format: date-time }
        rows:
          type: array
          items: { $ref: "#/components/schemas/TopicStockRow" }
    StockTopic:
      type: object
      description: "个股关联题材条目（eltdx helpers 风格）"
      properties:
        topic_id: { type: string }
        topic_name: { type: string }
        relation_level: { type: integer, nullable: true }
        selected_date: { type: string, nullable: true }
        topic_date: { type: string, nullable: true }
        reason: { type: string, nullable: true }
        detail_id: { type: string, nullable: true }
        category_raw: { type: string, nullable: true }
        source: { type: string, example: "helpers.stock_topics" }
        raw:
          type: object
          additionalProperties: true
    StockTopicsResponse:
      type: object
      description: "eltdx helpers.stock_topics 返回模型"
      properties:
        code: { type: string }
        count: { type: integer }
        source: { type: string, example: "eltdx" }
        fetched_at: { type: string, format: date-time }
        topics:
          type: array
          items: { $ref: "#/components/schemas/StockTopic" }

    # eltdx 行业/概念 指数 K 线
    IndexBar:
      type: object
      properties:
        time: { type: string, example: "2026-06-05 15:00:00+08:00" }
        open: { type: number }
        high: { type: number }
        low: { type: number }
        close: { type: number }
        prev_close:
          type: number
          nullable: true
          description: "上一根收盘价 (来自 eltdx last_close_price_milli / 1000)"
        pct:
          type: number
          nullable: true
          description: "涨跌幅 % = (close - prev_close) / prev_close * 100"
        volume_lots: { type: integer }
        amount: { type: number }
    IndexKlineResponse:
      type: object
      properties:
        code: { type: string, example: "sh880301" }
        name: { type: string, example: "农副食品" }
        kind: { type: string, enum: [industry, concept] }
        period: { type: string, example: "day" }
        count: { type: integer }
        source: { type: string, example: "eltdx" }
        bars:
          type: array
          items: { $ref: "#/components/schemas/IndexBar" }
    IndexCodeItem:
      type: object
      properties:
        code: { type: string, example: "sh880301" }
        name: { type: string, example: "农副食品" }
        kind: { type: string, enum: [industry, concept] }

    # ---- 行业 / 概念 应用面 ----
...